Fix: Migrate React app to Next.js
This commit addresses several issues related to migrating your React application to Next.js.

Changes include:
- Moved `supabaseClient.js` to `src/lib/` and updated all imports to use the new path (`@/lib/supabaseClient`).
- Migrated static `.jsx` pages in `src/pages/` to Next.js compatible `.js` pages, ensuring default exports.
- Replaced `<a>` tags with Next.js `<Link>` components for client-side navigation.
- Created `jsconfig.json` to support `@/*` alias paths.
- Created `.env.local` with placeholder Supabase credentials.

Note: I encountered environment issues during local build testing, so I could not verify the build locally.
