+++ src/components/modals/CreateEditTaskModal.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, useCallback } from 'react';
-import { supabase } from '../../supabaseClient';
+import { supabase } from '@/lib/supabaseClient';
 import { useAuth } from '../../context/AuthContext';
 
 // Helper function to generate a somewhat unique file name if needed
