+++ src/pages/staff.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState, useCallback } from 'react';
-import { supabase } from '../supabaseClient';
+import { supabase } from '@/lib/supabaseClient';
 import { useAuth } from '../context/AuthContext';
 import InviteStaffModal from '../components/modals/InviteStaffModal';
 import EditStaffModal from '../components/modals/EditStaffModal';
